New implementation based on porcelain status
A new function `git:status` now produces all the necessary information by
parsing the output of a single run of `git status --porcelain=v2`. This
reduces the number of git executions that need to be made. It returns a
map with the following keys (showing their default values):

```
  staged-modified = []
  staged-deleted  = []
  staged-added    = []
  local-modified  = []
  local-deleted   = []
  untracked       = []
  unmerged        = []
  ignored         = []
  renamed         = []
  copied          = []
  branch-name     = ""
  branch-oid      = ""
  rev-ahead       = 0
  rev-behind      = 0
  is-git-repo     = $false
```

List elements contain the list of files with that status. For each list
element, another one with the same key but ending in `-count` is
returned, with the number of files in that status. For example
`untracked` and `untracked-status`.

All the other functions are now wrappers around `status`. Note that
there is no caching, so if you call `git:branch_name` and
`git:rev_count`, it will result in two calls to `git`. I am not sure if
it's worth putting any logic for caching in this module, or leaving it
to the caller to keep usage simple. If you need caching, you can call
and store the output from `git:status` and just reuse the elements from
there.